fix: 매니저 params를 server components에서 사용 가능하도록 구현
