+++ apps/manager/app/league/[leagueId]/[gameId]/component.tsx
@@ -1,0 +1,130 @@
+'use client';
+import { useDeleteGame, useGame, useUpdateGame } from '@hcc/api';
+import { useToast } from '@hcc/ui';
+import { zodResolver } from '@hookform/resolvers/zod';
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+import { useForm } from 'react-hook-form';
+
+import {
+  gameDefaultValues,
+  GameForm,
+  gameFormSchema,
+  GameFormSchema,
+} from '@/app/league/[leagueId]/_components/GameForm';
+import AlertDialog from '@/components/AlertDialog';
+import Layout from '@/components/Layout';
+import { getStateByQuarter, QUARTER_KEY } from '@/constants/games';
+import { formatTime } from '@/utils/time';
+
+type ComponentProps = {
+  leagueId: string;
+  gameId: string;
+};
+
+const DeleteButton = ({ onAction }: { onAction: () => void }) => {
+  return (
+    <AlertDialog
+      title="삭제한 경기는 다시 복구할 수 없어요"
+      description="정말 삭제할까요?"
+      primaryActionLabel="삭제"
+      secondaryActionLabel="취소"
+      onPrimaryAction={onAction}
+    >
+      <button>경기 삭제</button>
+    </AlertDialog>
+  );
+};
+
+export const Component = ({ leagueId, gameId }: ComponentProps) => {
+  const router = useRouter();
+  const { toast } = useToast();
+
+  const { data: game } = useGame(gameId);
+
+  const methods = useForm<GameFormSchema>({
+    resolver: zodResolver(gameFormSchema),
+    defaultValues: gameDefaultValues,
+  });
+
+  useEffect(() => {
+    if (game) {
+      methods.reset({
+        name: game.gameName,
+        round: game.round.toString(),
+        quarter: game.gameQuarter,
+        startDate: new Date(formatTime(new Date(game.startTime), 'YYYY-MM-DD')),
+        startTime: formatTime(new Date(game.startTime), 'HH:mm'),
+        idOfTeam1: game.gameTeams[0]?.gameTeamId.toString() || '',
+        idOfTeam2: game.gameTeams[1]?.gameTeamId.toString() || '',
+        videoId: game.videoId,
+      });
+    }
+  }, [game, methods]);
+
+  const { mutate: updateGame, isPending } = useUpdateGame();
+  const onSubmit = (data: GameFormSchema) => {
+    if (isPending) return;
+
+    const quarter = data.quarter as QUARTER_KEY;
+    updateGame(
+      {
+        leagueId,
+        gameId,
+        name: data.name,
+        round: Number(data.round),
+        quarter: quarter,
+        state: getStateByQuarter(quarter),
+        startTime: `${formatTime(data.startDate, 'YYYY-MM-DD')}T${data.startTime}:00`,
+        videoId: data.videoId,
+      },
+      {
+        onSuccess: () => {
+          toast({
+            title: '경기 정보가 수정되었습니다.',
+            variant: 'destructive',
+          });
+          router.back();
+        },
+        onError: () => {
+          toast({
+            title: '경기 정보 수정에 실패했습니다.',
+            variant: 'destructive',
+          });
+        },
+      },
+    );
+  };
+
+  const { mutate: deleteGame } = useDeleteGame();
+  const handleDelete = () => {
+    deleteGame(
+      { leagueId, gameId },
+      {
+        onSuccess: () => {
+          toast({ title: '경기가 삭제되었습니다.', variant: 'destructive' });
+          router.replace(`/league/${leagueId}`);
+        },
+        onError: () => {
+          toast({ title: '경기 삭제에 실패했습니다.', variant: 'destructive' });
+        },
+      },
+    );
+  };
+
+  return (
+    <Layout
+      navigationTitle="경기 정보"
+      navigationMenu={<DeleteButton onAction={handleDelete} />}
+    >
+      <GameForm
+        leagueId={leagueId}
+        gameId={gameId}
+        methods={methods}
+        submitText="수정 완료"
+        onSubmit={onSubmit}
+        type="UPDATE"
+      />
+    </Layout>
+  );
+};
